requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
langchain-openai>=0.1.0
gradio>=6.0.0
openai>=1.0.0
python-dotenv>=1.0.0